реализована стандартная авторизация

+++ src/main/java/com/itc/controller/CourseController.java
@@ -2,14 +2,13 @@ package com.itc.controller;
 
 import com.itc.domain.Course;
 import com.itc.domain.User;
-import com.itc.exception.CourseNotFoundException;
-import com.itc.exception.UserNotFoundException;
 import com.itc.repos.CourseDao;
 import com.itc.repos.UserDao;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -40,9 +39,14 @@ public class CourseController {
     @PostMapping(value = "/api/admin/courses",
             consumes = "application/json",
             produces = "application/json")
-    public ResponseEntity<Course> createCourse(@RequestBody Course course) {
-        courseDao.save(course);
-        return new ResponseEntity<>(course, HttpStatus.OK);
+    public ResponseEntity<Course> createCourse(@RequestBody Course course,
+                                               @RequestBody User user) {
+        User userFromDb = userDao.findByEmail(user.getEmail());
+
+        if (userFromDb != null && userFromDb.isAuthentificated()) {
+            courseDao.save(course);
+            return new ResponseEntity<>(course, HttpStatus.OK);
+        } else return new ResponseEntity<>(course, HttpStatus.FORBIDDEN);
     }
 
     @PatchMapping(value = "/api/admin/courses/{id}",
@@ -50,27 +54,39 @@ public class CourseController {
             produces = "application/json")
     public ResponseEntity<Course> updateCourse(@PathVariable Long id,
                                                @RequestBody Course course,
+                                               @RequestBody User user,
                                                HttpServletResponse response) {
-        response.setHeader("Location", ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path("/findCourse/" + course.getId()).toUriString());
-        Course courseFromDB = courseDao.findById(id).get();
-        BeanUtils.copyProperties(course, courseFromDB);
-        courseDao.save(courseFromDB);
-        return new ResponseEntity<>(courseFromDB, HttpStatus.OK);
+        User userFromDb = userDao.findByEmail(user.getEmail());
+
+        if (userFromDb != null && userFromDb.isAuthentificated()) {
+            response.setHeader("Location", ServletUriComponentsBuilder.fromCurrentContextPath()
+                    .path("/findCourse/" + course.getId()).toUriString());
+            Course courseFromDB = courseDao.findById(id).get();
+            BeanUtils.copyProperties(course, courseFromDB);
+            courseDao.save(courseFromDB);
+
+            return new ResponseEntity<>(courseFromDB, HttpStatus.OK);
+        } else return new ResponseEntity<>(course, HttpStatus.FORBIDDEN);
     }
 
     @DeleteMapping(value = "/api/admin/courses/{id}",
             consumes = "application/json",
             produces = "application/json")
-    public ResponseEntity<Course> deleteCourse(@PathVariable Long id, HttpServletResponse response) {
-        response.setHeader("Location", ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path("/findCourse/" + id).toUriString());
-        courseDao.deleteById(id);
+    public ResponseEntity<Course> deleteCourse(@PathVariable Long id,
+                                               HttpServletResponse response,
+                                               @RequestBody User user) {
+        User userFromDb = userDao.findByEmail(user.getEmail());
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (userFromDb != null && userFromDb.isAuthentificated()) {
+            response.setHeader("Location", ServletUriComponentsBuilder.fromCurrentContextPath()
+                    .path("/findCourse/" + id).toUriString());
+            courseDao.deleteById(id);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PatchMapping(value = "/api/admin/courses/{courseId}/enroll",
+ /*   @PatchMapping(value = "/api/admin/courses/{courseId}/enroll",
             consumes = "application/json",
             produces = "application/json")
     public ResponseEntity<Course> enrollOnCourse(@PathVariable Long courseId,
@@ -87,11 +103,11 @@ public class CourseController {
             if(userFromDb == null) {
                 throw new UserNotFoundException();
             }
-            courseFromDB.setStudents(courseFromDB.getStudents().add(userFromDb));
+         //   courseFromDB.setStudents(courseFromDB.getStudents().add(userFromDb));
             courseDao.save(courseFromDB);
         } catch (CourseNotFoundException | UserNotFoundException) {
             return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
         return new ResponseEntity<>(HttpStatus.OK);
-    }
+    }*/
 }
